chore: Use v4_10 gtk4-rs feature
Gtk.FileDialog allows selecting from directories that are not explicitly
permitted within the Flatpak manifest, and grants the app permission
upon a user doing so. Therefore, `is_accessible` and the related dialog
are no longer relevant. Do note that if not explicitly allowed, the
directory will show as a not-so-pretty sandboxed path.